// react/src/api/models/acroStack/services/dtos/cms/CommentDto.ts
/* oxlint-disable */

/**
 * @type object
 */
export type AcroStackServicesDtosCmsCommentDto = {
  /**
   * @description
   * Format: `uuid`
   * @type string | undefined
   */
  id?: string;
  /**
   * @description
   * Format: `date-time`
   * @type string | undefined
   */
  creationTime?: string;
  /**
   * @description
   * Format: `uuid`
   * @type string | undefined
   */
  creatorId?: string | null;
  /**
   * @description
   * Format: `date-time`
   * @type string | undefined
   */
  lastModificationTime?: string | null;
  /**
   * @description
   * Format: `uuid`
   * @type string | undefined
   */
  lastModifierId?: string | null;
  /**
   * @type boolean | undefined
   */
  isDeleted?: boolean;
  /**
   * @description
   * Format: `uuid`
   * @type string | undefined
   */
  deleterId?: string | null;
  /**
   * @description
   * Format: `date-time`
   * @type string | undefined
   */
  deletionTime?: string | null;
  /**
   * @type string | undefined
   */
  entityType?: string | null;
  /**
   * @description
   * Format: `uuid`
   * @type string | undefined
   */
  entityId?: string;
  /**
   * @type string | undefined
   */
  text?: string | null;
  /**
   * @type string | undefined
   */
  authorName?: string | null;
  /**
   * @description
   * Format: `uuid`
   * @type string | undefined
   */
  parentId?: string | null;
  /**
   * @description
   * Format: `uuid`
   * @type string | undefined
   */
  creatorUserId?: string | null;
  /**
   * @type string | undefined
   */
  creatorUserName?: string | null;
};